+++ irida_staramr_results/downloader.py
@@ -158,7 +158,7 @@ def _files_to_data_frames(results_files):
     """
     data_frames = {}
     for file in results_files:
-        data_frames[file.get_sheet_name()] = _convert_to_df(file.get_contents())
+        data_frames[file.get_sheet_name()] = _convert_to_df(file.get_sheet_name(), file.get_contents())
 
     return data_frames
 
@@ -177,24 +177,29 @@ def _append_file_data_to_existing_data_frames(results_files, data_frames):
         file_sheet_name = file.get_sheet_name()
         if file_sheet_name not in data_frames.keys():
             # new file, new dataframe
-            data_frames[file_sheet_name] = _convert_to_df(file.get_contents())
+            data_frames[file_sheet_name] = _convert_to_df(file_sheet_name, file.get_contents())
         else:
             # appending data to existing dataframe
             prev_data = data_frames[file_sheet_name]
-            curr_data = _convert_to_df(file.get_contents())
+            curr_data = _convert_to_df(file_sheet_name, file.get_contents())
             updated_data = prev_data.append(curr_data)
             data_frames[file_sheet_name] = updated_data
 
     return data_frames
 
 
-def _convert_to_df(file_content):
+def _convert_to_df(file_sheet_name, file_content):
     """
     Converts dictionary or tsv contents to a data frame
+    :param file_sheet_name: Used for checking if data is being pulled from an excel sheet in the case of PointFinder
     :param file_content:
     :return data_frame:
     """
-    if type(file_content) is dict:
+    # Pointfinder data comes from the specified "PointFinder" page of an excel sheet.
+    if file_sheet_name == "PointFinder":
+        data_frame = pd.read_excel(file_content, sheet_name=file_sheet_name)
+    # All other data fits into either a dict or a csv/tsv
+    elif type(file_content) is dict:
         data_frame = pd.DataFrame([file_content])
     else:
         data_frame = pd.read_csv(io.StringIO(file_content), delimiter="\t")