+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changes
 
+##0.2.0 to 0.3.0
+Bug Fixes
+* Fixed pointfinder data not being added to outputs correctly
+
+Developer Changes
+* Added `openpyxl` as a dependency for opening excel files with `pandas`
+
 ##0.1.2 to 0.2.0
 
 Features